Working auto, with test-bot coefficients, for all four starting positions (+updated testbot coefficients for DriveTeleop)

+++ TeamCode/src/main/java/com/pocolifo/robobase/StartSide.java
@@ -5,7 +5,7 @@ public enum StartSide {
     BACKDROP_SIDE;
 
 
-    public int getSideSwapConstantIn() {
+    public double getSideSwapConstantIn() {
         return this == StartSide.APRIL_TAG_SIDE ? 48 : 0;
     }
 }
